Crate librespot_audio

Structs

AudioDecrypt

StreamLoaderController

Enums

AudioFile

AudioFileError

Constants

MINIMUM_DOWNLOAD_SIZE

The minimum size of a block that is requested from the Spotify servers in one request. This is the block size that is typically requested while doing a seek() on a file. The Symphonia decoder requires this to be a power of 2 and > 32 kB. Note: smaller requests can happen if part of the block is downloaded already.

READ_AHEAD_BEFORE_PLAYBACK

Before playback starts, this many seconds of data must be present. Note: the calculations are done using the nominal bitrate of the file. The actual amount of audio data may be larger or smaller.

READ_AHEAD_BEFORE_PLAYBACK_ROUNDTRIPS

Same as READ_AHEAD_BEFORE_PLAYBACK, but the time is taken as a factor of the ping time to the Spotify server. Both READ_AHEAD_BEFORE_PLAYBACK and READ_AHEAD_BEFORE_PLAYBACK_ROUNDTRIPS are obeyed. Note: the calculations are done using the nominal bitrate of the file. The actual amount of audio data may be larger or smaller.

READ_AHEAD_DURING_PLAYBACK

While playing back, this many seconds of data ahead of the current read position are requested. Note: the calculations are done using the nominal bitrate of the file. The actual amount of audio data may be larger or smaller.

READ_AHEAD_DURING_PLAYBACK_ROUNDTRIPS

Same as READ_AHEAD_DURING_PLAYBACK, but the time is taken as a factor of the ping time to the Spotify server. Note: the calculations are done using the nominal bitrate of the file. The actual amount of audio data may be larger or smaller.